ledger/blockstore: Use `u32` for shred indices.
Shred indices, within a slot, can not exceed a `u32` value.
Unfortunately, there is no common type that is used when a shred index
needs to be stored.  Part of the code use `u32`, another considerable
part uses `u64` and a few places use `usize`.

Ideally, it would be better to have a dedicated type, wrapping a
primitive `u32` value.

This change is trying to unify the types used, moving most of the code
that works with shred indices to `u32`.  `u64` is used in a few
locations that are serialized into the blockstore or are part of the
wire protocol.  Those are left as is for now.

As the conversion is not 100%, there are still a few conversions left,
that could be avoided should both the serialization and the wire
protocol code use `u32`.  I also was not completely sure about the
`usize` types used by the Merkel tree code - those might be convertible
to `u32` as well.

Stats (roughly):

  * Removed `as u{32,64}`:

  $ git diff --cached | grep '^-.*as u\(32\|64\)' | wc -l
  54

  * Removed `unwrap()`/`expect()`:

  $ git diff --cached | grep '^-.*\(unwrap()\|expect()\)' | wc -l
  18

  * Added `as u{32,64}`:

  $ git diff --cached | grep '^+.*as u\(32\|64\)' | wc -l
  8

  * Added `unwrap()`/`expect()`:

  $ git diff --cached | grep '^+.*\(unwrap()\|expect()\)' | wc -l
  31

Total is 54 + 18 - 8 - 31 = 33 conversions removed.  And a lot of the
removed conversion are unchecked, while a lot of the added ones are
checked.  Current version would still panic, if a value outside the
`u32` scope is used - it will just panic further away from the bug
location, complicating the debugging efforts.

This change also prepares the grounds for removal of most of the
conversion that are still present.  We would need to convert the
serialization and the wire code to `u32` as well.

There are a number of `as usize` conversions added in the new code.
Those should be safe, as they convert from `u32`.  Most of them are used
when indexing into `Vec`s.

+++ core/src/progress_map.rs
@@ -41,7 +41,7 @@ impl ReplaySlotStats {
         slot: Slot,
         num_txs: usize,
         num_entries: usize,
-        num_shreds: u64,
+        num_shreds: u32,
         bank_complete_time_us: u64,
     ) {
         lazy! {
